add doc, add findbugs exclusion, and move bench project position
